+++ NEWS.md
@@ -5,9 +5,14 @@
  * Now `availableCores()` recognizes environment variable
    `IS_BIOC_BUILD_MACHINE`, which is set to true by the Bioconductor
    (>= 3.16) check servers.  If true, then a maximum of four (4) cores
-   is returned, regardless of environment variable
-   `BIOCPARALLEL_WORKER_NUMBER`.  This new environment variable
-   replaces legacy variable `BBS_HOME` used in Bioconductor (<= 3.15).
+   is returned.  This new environment variable replaces legacy
+   variable `BBS_HOME` used in Bioconductor (<= 3.15).
+   
+ * `availableCores()` split up method `"BiocParallel"` into two;
+   `"BiocParallel"` and `"Bioconductor"`.  The former queries
+   environment variable `BIOCPARALLEL_WORKER_NUMBER` and the latter
+   `IS_BIOC_BUILD_MACHINE`.  This means `availableCores(which =
+   "all")` now reports on both.
 
 ## Documentation